fix ci breaks caused by implicit conversion loses

+++ onnxruntime/core/providers/webgpu/math/einsum.h
@@ -10,13 +10,16 @@
 namespace onnxruntime {
 namespace webgpu {
 struct SymbolInfo {
-  int count{0};
+  size_t count{0};
   std::vector<int> input_indices;
-  int dim_value{0};
+  int64_t dim_value{0};
 };
 
 struct EinsumTerm {
-  std::map<std::string, std::vector<int>> symbol_to_indices;
+  // Indices of the symbols in the term which cannot be negative.
+  std::map<std::string, std::vector<size_t>> symbol_to_indices;
+  // The index of the input tensor in the Einsum equation.
+  // This is -1 for the output term.
   int input_index{-1};
 };
 
@@ -31,7 +34,7 @@ class EinsumEquation {
  private:
   bool has_ellipsis_{false};
   std::vector<int64_t> ellipsis_dims_;
-  void AddSymbol(const std::string& symbol, int dim_value, int input_index);
+  void AddSymbol(const std::string& symbol, int64_t dim_value, int input_index);
   EinsumTerm ProcessTerm(const std::string& term,
                          bool is_input,
                          gsl::span<const int64_t> dims,
@@ -40,7 +43,7 @@ class EinsumEquation {
 
 class EinsumProgram final : public Program<EinsumProgram> {
  public:
-  EinsumProgram(int input_count, const EinsumEquation& parsed_equation)
+  EinsumProgram(size_t input_count, const EinsumEquation& parsed_equation)
       : Program{"Einsum"}, input_count_(input_count), parsed_equation_{parsed_equation} {}
 
   Status GenerateShaderCode(ShaderHelper& sh) const override;
@@ -48,7 +51,7 @@ class EinsumProgram final : public Program<EinsumProgram> {
   WEBGPU_PROGRAM_DEFINE_UNIFORM_VARIABLES({"output_size", ProgramUniformVariableDataType::Uint32});
 
  private:
-  int input_count_;
+  size_t input_count_;
   const EinsumEquation& parsed_equation_;
 };
